archive/lambda_archive/requirements.txt
python-telegram-bot
aiohttp
deepgram-sdk
python-dotenv
requests
google.generativeai
